requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.10.4
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
elasticsearch==8.17.0
fastapi==0.110.1
google-api-core==2.24.0
google-api-python-client==2.125.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
protobuf==5.29.2
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.29.0